# internal/openapi/swagger.yaml
basePath: /api
definitions:
  api.APIEnvelope:
    properties:
      data:
        type: object
      message:
        example: ok
        type: string
      success:
        example: true
        type: boolean
    type: object
  api.APIErrorEnvelope:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: true
        type: boolean
      message:
        example: invalid request
        type: string
    type: object
...
  api.LogEntryDTO:
    properties:
      action:
        example: connect
        type: string
      group:
        example: tunnel
        type: string
      level:
        example: info
        type: string
      message:
        example: Tunnel connected
        type: string
      subgroup:
        example: tun_abc123
        type: string
      target:
        example: vpn.example.com:51820
        type: string
      timestamp:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
  api.LogsData:
    properties:
      bucket:
        example: app
        type: string
      bufferCapacity:
        example: 5000
        type: integer
      bufferSize:
        example: 123
        type: integer
      enabled:
        example: true
        type: boolean
      logs:
        items:
          $ref: '#/definitions/api.LogEntryDTO'
        type: array
      oldestTimestamp:
        example: "2024-01-15T08:00:00Z"
        type: string
      total:
        example: 42
        type: integer
    type: object
  api.LogsResponseEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.LogsData'
      success:
        example: true
        type: boolean
    type: object
...
  api.SubgroupsData:
    properties:
      group:
        example: routing
        type: string
      subgroups:
        items:
          type: string
        type: array
    type: object
  api.SubgroupsResponseEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.SubgroupsData'
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /logs:
    get:
      description: Returns log entries from the selected bucket. `bucket=app` (default)
        covers tunnel/routing/server/system events; `bucket=singbox` covers sing-box
        forwarder events isolated from app history.
      parameters:
      - description: Bucket selector
        enum:
        - app
        - singbox
        in: query
        name: bucket
        type: string
      - collectionFormat: multi
        description: Filter by group (repeat param or comma-separated values)
        in: query
        items:
          type: string
        name: group
        type: array
      - collectionFormat: multi
        description: Filter by subgroup (repeat param or comma-separated values)
        in: query
        items:
          type: string
        name: subgroup
        type: array
      - description: Filter by level
        in: query
        name: level
        type: string
      - description: Max entries to return (default 200)
        in: query
        name: limit
        type: integer
      - description: Skip first N matching entries
        in: query
        name: offset
        type: integer
      - description: Unix seconds — return only entries strictly after this
        in: query
        name: since
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LogsResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Get logs
      tags:
      - logs
  /logs/clear:
    post:
      description: Clears all entries from the requested bucket. `bucket` is required
        — there is no implicit "clear everything" since app and sing-box logs serve
        different audiences.
      parameters:
      - description: Bucket to clear
        enum:
        - app
        - singbox
        in: query
        name: bucket
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Clear logs
      tags:
      - logs
  /logs/subgroups:
    get:
      description: Returns the static subgroup catalog from internal/logging.KnownSubgroups.
        Order is presentation-stable. Empty group returns 400.
      parameters:
      - description: Group name
        enum:
        - tunnel
        - routing
        - server
        - system
        - singbox
        in: query
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubgroupsResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: List known subgroups for a group
      tags:
      - logs
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"